# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stellar Insights API",
    "description": "API for Stellar network analytics, anchor monitoring, and payment corridor insights",
    "contact": {
      "name": "Stellar Insights Team",
      "email": "support@stellarinsights.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://api.stellarinsights.io",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/alerts/rules/{rule_id}": {
      "put": {
        "tags": ["Alerts"],
        "summary": "Update an alert rule",
        "operationId": "update_rule",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Alert rule updated" }
        }
      },
      "delete": {
        "tags": ["Alerts"],
        "summary": "Delete an alert rule",
        "operationId": "delete_rule",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Alert rule deleted" }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}